backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "datetime>=5.5",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "motor>=3.7.1",
    "nltk>=3.9.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.0",
    "spacy>=3.8.7",
    "pydantic-settings>=2.0.0",
    "pip>=25.1.1",
]

[dependency-groups]
dev = [
    "uvicorn>=0.34.3",
]
